% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FL.R
\name{FL}
\alias{FL}
\title{FL: uses local summaries to fit a linear regression model}
\usage{
FL(summary)
}
\arguments{
\item{summary}{a list containing the combined local summaries from individual servers.}
}
\value{
A dataframe made up of the foloowing columns
\item{Estimate}{The estimated values of the regression coefficients}
\item{Std.Error}{The standard errors of the estimates}
\item{t value}{t values computed using the estimates and standard errors}
\item{p value}{p values of the t statistics used to test the significance of the regression coefficients}
}
\description{
This comprehensive function implements federated learning procedure by combining summaries fro local servers to fit a linear model.
It provides estimates of the Regression coefficients, standard errors, t values, p values of the estimates.
}
\examples{
set.seed(42)  # For reproducibility
# Simulating dataset1
age1 = rnorm(100, mean = 50, sd = 10)  # simulating age
income1 = rnorm(100, mean = 50000, sd = 10000)  # simulating income
noise1 = rnorm(100, 0, 1)  # some noise
Purchase1 = 20000 + 500*age1 + 0.05*income1 + 10*age1*income1/100000 + noise1  # creating a relationship for Purchase
data1 = data.frame(age = age1, income = income1, Purchase = Purchase1)
# Simulating dataset2 in a similar way
age2 = rnorm(100, mean = 50, sd = 10)
income2 = rnorm(100, mean = 50000, sd = 10000)
noise2 = rnorm(100, 0, 1)
Purchase2 = 20000 + 500*age2 + 0.05*income2 + 10*age2*income2/100000 + noise2
data2 = data.frame(age = age2, income = income2, Purchase = Purchase2)
summary1 = FL_local_summary(Purchase ~ age + income +income:age, data1)
summary2 = FL_local_summary(Purchase ~ age + income + income:age, data2)
combined_summaries = FL_combine(list(summary1, summary2))
soln = FL(combined_summaries)
print(soln)

}
\author{
Justice Akuoko-Frimpong

Jonathan Ta

Michael Kaye
}
