% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FL.R
\name{FL_combine}
\alias{FL_combine}
\title{FL_combine: computes the combine summary}
\usage{
FL_combine(summaries)
}
\arguments{
\item{summaries}{A list containing local summaries from individual groups}
}
\value{
A list including elements
\item{xtx}{a matrix for the combined sum of squares of X}
\item{xty}{a matrix for the combined sum of squares of XY}
\item{yty}{a scalar value for the combined sum of squares of Y}
\item{n}{the total sample size}
}
\description{
This function computes the combine summary using individual local summaries.
}
\details{
Description of Function 2.
}
\examples{
set.seed(42)  # For reproducibility
# Simulating dataset1
age1 = rnorm(100, mean = 50, sd = 10)  # simulating age
income1 = rnorm(100, mean = 50000, sd = 10000)  # simulating income
noise1 = rnorm(100, 0, 1)  # some noise
Purchase1 = 20000 + 500*age1 + 0.05*income1 + 10*age1*income1/100000 + noise1  # creating a relationship for Purchase
data1 = data.frame(age = age1, income = income1, Purchase = Purchase1)
summary1 = FL_local_summary(Purchase ~ age + income +income:age, data1)
print(summary1)
# Simulating dataset2 in a similar way
age2 = rnorm(100, mean = 50, sd = 10)
income2 = rnorm(100, mean = 50000, sd = 10000)
noise2 = rnorm(100, 0, 1)
Purchase2 = 20000 + 500*age2 + 0.05*income2 + 10*age2*income2/100000 + noise2
data2 = data.frame(age = age2, income = income2, Purchase = Purchase2)
summary2 = FL_local_summary(Purchase ~ age + income +income:age, data2)
combined_summaries = FL_combine(list(summary1, summary2))
}
