% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghrsst_daily.R
\name{ghrsst_daily_sst_files}
\alias{ghrsst_daily_sst_files}
\title{an experiment with raadtools sst}
\usage{
ghrsst_daily_sst_files(date)
}
\arguments{
\item{date}{the dates you want}
}
\value{
files, data frame with columns date, datasource
}
\description{
date in put is processed by raad.times
}
