% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_daily.R
\name{oisst_daily_sst_files}
\alias{oisst_daily_sst_files}
\title{an experiment with raadtools sst}
\usage{
oisst_daily_sst_files(date)
}
\arguments{
\item{date}{the dates you want}
}
\value{
files, data frame with columns date, datasource
}
\description{
date in put is processed by raad.times
}
\examples{
oisst_daily_sst_files(as.Date("2002-01-01") + sort(sample(1:100, 20)))
}
